📌 ISSUE-#30: Improving unit tests
