+++ samples/simple_function_workflow.py
@@ -14,8 +14,7 @@ def main():
     workflow.task.add(step=simple_step)
     workflow.start()
 
-    for task in workflow.result_task():
-        print(task.task_id, task.status, task.current_context.storage)
+    return workflow
 
 
 if __name__ == "__main__":
