+++ samples/workflow_retry.py
@@ -15,6 +15,8 @@ def main():
     workflow.task.add(step=simple_step)
     workflow.start()
 
+    return workflow
+
 
 if __name__ == "__main__":
     main()
